Add files via upload
This is a landlord and tenant money management software. Some more features will be added here, such that the tenant will be able to see his Duke rent or bill through his login and will be able to make payments online. Also, notifications will be sent to the landlord and the tenant on various issues. The tenant can complain if he wants. Also, I want to provide a login feature for the co-admin or caretaker. Let me see what can be done. Well wish for me.